backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# LangChain
langchain==0.3.7
langchain-groq==0.2.1
langchain-chroma==0.1.4
langchain-huggingface==0.1.2
langchain-core==0.3.15

# Vector DB and Search
chromadb==0.5.20
rank-bm25==0.2.2

# LandingAI ADE
landingai-ade==1.5.0

# Utilities
numpy==1.26.4
pydantic==2.10.3
